Use points instead of pixels for sizes

+++ src/ShapeCrawler/Positions/IPosition.cs
@@ -8,12 +8,12 @@ namespace ShapeCrawler;
 public interface IPosition
 {
     /// <summary>
-    ///     Gets or sets x-coordinate of the upper-left corner of the shape.
+    ///     Gets or sets x-coordinate of the upper-left corner of the shape in points.
     /// </summary>
-    decimal X { get; set; }
+    float X { get; set; }
 
     /// <summary>
-    ///     Gets or sets y-coordinate of the upper-left corner of the shape in pixels.
+    ///     Gets or sets y-coordinate of the upper-left corner of the shape in points.
     /// </summary>
-    decimal Y { get; set; }
+    float Y { get; set; }
 }
